[CHANGE] Show "X" for DNR items and increase the blur strength
The "X" feels as an important way to carry the meaning that the item is a limit, in similar way to the "prevent".
The blur was strengthened because of how we think about the feature:
its purpose is to give people that are disturbed by specific item a tool to avoid seeing it under any circumstances.
For that reason we think that the preview should be blurred past being recognizable.