[CHANGE] Allow changing color of items by default and change permission name
This is because it is common to add item on other characters and then color it.
Also nothing is preventing you from removing the item, coloring it, and adding it back on.

This permission is therefore meant more for players that have problems with items changing color while on them.
We, however, think that for majority of players it is better to allow it, as otherwise it would be simply inconvenience.