[FIX] Start database process detached to prevent killing it too early
The way terminal works is, that if user does Ctrl+C in the terminal, the signal is sent to the whole _process group_.
This means that Mongo was receiving the signal too, and shutting down before Pandora itself had chance to.
By detaching it this doesn't happen.

Also note, that the `mongodb-memory-server` has a killer script that periodically checks if parent process is still running, so the database will get killed even if the main process dies on the spot, even if detached.